css/css-flexbox/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS Flexible Box Layout properties</title>
<link rel="help" href="https://drafts.csswg.org/css-flexbox-1/#property-index">
<meta name="assert" content="Properties inherit or not according to the spec.">
<meta name="assert" content="Properties have initial values according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
  <div id="target"></div>
</div>
<style>
  #container, #target {
    display: flex;
  }
</style>
<script>
assert_not_inherited('align-content', 'stretch', 'center'); // expected "stretch" but got "normal"
assert_not_inherited('align-items', 'stretch', 'center'); // expected "stretch" but got "normal"
assert_not_inherited('align-self', 'auto', 'stretch');
assert_not_inherited('flex-basis', 'auto', '10px');
assert_not_inherited('flex-direction', 'row', 'column');
assert_not_inherited('flex-grow', '0', '2');
assert_not_inherited('flex-shrink', '1', '3');
assert_not_inherited('flex-wrap', 'nowrap', 'wrap-reverse');
assert_not_inherited('justify-content', 'flex-start', 'center'); // expected "flex-start" but got "normal"
assert_not_inherited('order', '0', '2');
</script>
</body>
</html>
